+++ SharpSqlBuilder/Extensions/TypeExtensions.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
 using SharpSqlBuilder.Attributes;
 using SharpSqlBuilder.Maps;